Add option to pass labelpad argument to x/ylabel
By default, the code uses fixed coordinates for the x and y axes labels. This sometimes interferes with the ticklabels if the values have a lot of digits.
This change allows one to pass the labelpad argument to the set_xlabel or set_ylabel commands to better control the label spacing when needed.
The default behavior is not changed.