Allow the user to call tight_layout
The use of `set_visible(False)` breaks the `tight_layout()` routine and
produces a AttributeError when it is called.  This replaces the
`set_visible` call by removing the ticks manually and, since there is
now no need to force the `tight_layout` to be False, also removes the
occurances of this.