Support passing an empty Figure as the `fig` kwarg.
Small helper to support passing a new, empty Figure as the `fig` kwarg
(useful e.g. when embedding Matplotlib in a GUI -- the empty Figure
object may be created as part of the GUI setup).

Also document the supported values for `fig` -- in particular that if it
is unset, a *new* figure will be created, rather than overplotting the
current figure (which is a quite common behavior for libraries built
over Matplotlib, even though it doesn't really make sense here).

Matplotlib minimum version was bumped to 2.1 (2017) as that's what added
`Figure.subplots`.